+++ iosApp/Coffeegram/Coffeegram/CoffeeDrink.swift
@@ -7,7 +7,7 @@ final class CoffeeDrink: Identifiable {
     let name: String
     let icon: String // SF Symbol name
     var count: Int
-    
+
     init(name: String, icon: String, count: Int) {
         self.name = name
         self.icon = icon
